requirements.txt
pandas>=2.0
numpy>=1.24
matplotlib>=3.7
seaborn>=0.12
plotly>=5.0
pyyaml>=6.0
scikit-learn>=1.3
xgboost>=2.0
tqdm>=4.65
